fix some pedantic compiler warnings
this allows the test suite of eWoms to be compiled on clang++ 3.8
using the following warning flags

```
    -Weverything
    -Wno-documentation
    -Wno-documentation-unknown-command
    -Wno-c++98-compat
    -Wno-c++98-compat-pedantic
    -Wno-undef
    -Wno-padded
    -Wno-global-constructors
    -Wno-exit-time-destructors
    -Wno-weak-vtables
    -Wno-float-equal
```

without triggering warnings in opm-material or eWoms.

Note that the test-suite for opm-material does *not* yet pass without
warnings with these flags because with these flags clang likes to
complain about reducing or increasing floating point precision which
results in a formidable nightmare. I will see what I can do about
these warnings in a separate PR.

v4: properly fix the unused parameter warnings for the valgrind client
requests. thanks to [at]atgeirr for finding the issue.

+++ opm/material/fluidstates/FluidStateFugacityModules.hpp
@@ -43,8 +43,8 @@ namespace Opm {
  *        phase fugacity coefficients explicitly.
  */
 template <class Scalar,
-          int numPhases,
-          int numComponents,
+          unsigned numPhases,
+          unsigned numComponents,
           class Implementation>
 class FluidStateExplicitFugacityModule
 {
@@ -111,12 +111,12 @@ protected:
  *        fugacity coefficients explicitly assuming immiscibility.
  */
 template <class Scalar,
-          int numPhases,
-          int numComponents,
+          unsigned numPhases,
+          unsigned numComponents,
           class Implementation>
 class FluidStateImmiscibleFugacityModule
 {
-    static_assert((int) numPhases == (int) numComponents,
+    static_assert(numPhases == numComponents,
                   "The number of phases must be the same as the number of (pseudo-) components if you assume immiscibility");
 
 public:
@@ -187,13 +187,13 @@ public:
     /*!
      * \brief The fugacity coefficient of a component in a phase []
      */
-    const Scalar& fugacityCoefficient(int /* phaseIdx */, int /* compIdx */) const
+    const Scalar& fugacityCoefficient(unsigned /* phaseIdx */, unsigned /* compIdx */) const
     { OPM_THROW(std::logic_error, "Fugacity coefficients are not provided by this fluid state"); }
 
     /*!
      * \brief The fugacity of a component in a phase [Pa]
      */
-    const Scalar& fugacity(int /* phaseIdx */, int /* compIdx */) const
+    const Scalar& fugacity(unsigned /* phaseIdx */, unsigned /* compIdx */) const
     { OPM_THROW(std::logic_error, "Fugacities coefficients are not provided by this fluid state"); }
 
     /*!
